Modify: D/L constructor to match class

+++ src/app/src/main/java/com/hcmus/group14/moneytor/data/model/DebtLend.java
@@ -48,7 +48,7 @@ public class DebtLend implements Serializable {
         this("", 0, 0, 0, 0L, "");
     }
 
-    public DebtLend(@Nullable String category, int value, int target,
+    public DebtLend(@Nullable String category, long value, int target,
                     int debt, long date, @Nullable String desc)
     {
         assert(debt == 0 || debt == 1);
